Initial support for links under Windows
Tested only with 'link' (not 'link-target').
For now, does all the linking under C:/Program Files/yotta/...
